+++ CHANGELOG.rst
@@ -8,7 +8,10 @@ All notable changes to this project are documented in this file.
 - Implementing interop type ``MAP`` along with new opcodes ``NEWMAP HASKEY KEYS VALUES`` and modify ``ARRAYSIZE PICKITEM SETITEM REMOVE`` to support ``MAP`` as `per PR here <https://github.com/neo-project/neo-vm/pull/28>__`
 - Added support for using ``--from-addr=`` to specify the address to use for ``testinvoke`` in ``prompt.py``. (`PR #329 <https://github.com/CityOfZion/neo-python/pull/329>`_)
 - Added TransactionInvocation.GetScript to ``StateReader.py``
-- Fixed missing uri locations in ``neo/api/REST/RestApi.py``
+- Fixed missing uri locations in ``neo/api/REST/RestApi.py`` (`PR #342 <https://github.com/CityOfZion/neo-python/pull/342>`_)
+- Fixed privatenet check by fixing the chain path for checks in Settings (`PR #341 <https://github.com/CityOfZion/neo-python/pull/341>`_)
+- Fixed ``neo-privnet.sample.wallet``
+
 
 [0.6.1] 2018-03-16
 ----------------------------
@@ -25,7 +28,7 @@ All notable changes to this project are documented in this file.
     - ``np-reencrypt-wallet``
   - updated docs for Pypi changes
 
-      
+
 [0.5.7] 2018-03-14
 ----------------------------
 - update to ``neocore==0.3.8``